pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ybi-strategy"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "requests>=2.31",
  "PyYAML>=6.0",
  "python-dateutil>=2.8",
  "tzdata>=2023.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

